+++ cmd/config.go
@@ -3,8 +3,9 @@ package cmd
 import "github.com/spf13/cobra"
 
 var configCmd = &cobra.Command{
-	Use:   "config",
-	Short: "Manage clerk configuration",
+	Use:               "config",
+	Short:             "Manage clerk configuration",
+	ValidArgsFunction: noFileComp,
 	RunE: func(cmd *cobra.Command, args []string) error {
 		return configShowCmd.RunE(cmd, args)
 	},